+++ lib/selectDialog.dart
@@ -1,5 +1,6 @@
-import 'package:flutter/material.dart';
 import 'dart:async';
+
+import 'package:flutter/material.dart';
 
 typedef Widget SelectOneItemBuilderType<T>(
     BuildContext context, T item, bool isSelected);
@@ -8,59 +9,71 @@ class SelectDialog<T> extends StatefulWidget {
   final T selectedValue;
   final List<T> itemsList;
   final bool showSearchBox;
-  final bool isFilterOnline;
+  final bool isFilteredOnline;
   final void Function(T) onChange;
   final Future<List<T>> Function(String text) onFind;
   final SelectOneItemBuilderType<T> itemBuilder;
   final InputDecoration searchBoxDecoration;
   final Color backgroundColor;
-  final TextStyle titleStyle;
+  final TextStyle dialogTitleStyle;
   final String Function(T item) itemAsString;
+  final bool Function(T item, String filter) filterFn;
   final String hintText;
+  final bool isMenuMode;
+  final double maxHeight;
+  final String dialogTitle;
 
   const SelectDialog(
       {Key key,
+      this.dialogTitle,
       this.itemsList,
+      this.maxHeight,
+      this.isMenuMode,
       this.showSearchBox,
-      this.isFilterOnline,
+      this.isFilteredOnline,
       this.onChange,
       this.selectedValue,
       this.onFind,
       this.itemBuilder,
       this.searchBoxDecoration,
-      this.backgroundColor = Colors.white,
-      this.titleStyle,
+      this.backgroundColor,
+      this.dialogTitleStyle,
       this.hintText,
-      this.itemAsString})
+      this.itemAsString,
+      this.filterFn})
       : super(key: key);
 
   static Future<T> showModal<T>(BuildContext context,
       {List<T> items,
       String label,
-      bool isFilteredOnline = false,
+      String dialogTitle,
+      isMenuMode,
+      double maxHeight,
+      bool isFilteredOnline,
       T selectedValue,
       bool showSearchBox,
       Future<List<T>> Function(String text) onFind,
       String Function(T item) itemAsString,
+      bool Function(T item, String filter) filterFn,
       SelectOneItemBuilderType<T> itemBuilder,
       void Function(T) onChange,
       InputDecoration searchBoxDecoration,
       Color backgroundColor,
       String hintText,
-      TextStyle titleStyle}) {
+      TextStyle dialogTitleStyle}) {
     return showDialog(
       context: context,
       builder: (context) {
         return AlertDialog(
           backgroundColor: backgroundColor,
-          title: Text(
-            label ?? "",
-            style: titleStyle,
-          ),
           content: SelectDialog<T>(
+              isMenuMode: isMenuMode,
+              dialogTitle: dialogTitle,
+              maxHeight: maxHeight,
+              isFilteredOnline: isFilteredOnline,
               itemAsString: itemAsString,
+              filterFn: filterFn,
               selectedValue: selectedValue,
-              isFilterOnline: isFilteredOnline,
               itemsList: items,
               onChange: onChange,
               onFind: onFind,
@@ -69,10 +82,62 @@ class SelectDialog<T> extends StatefulWidget {
               searchBoxDecoration: searchBoxDecoration,
               backgroundColor: backgroundColor,
               hintText: hintText,
-              titleStyle: titleStyle),
+              dialogTitleStyle: dialogTitleStyle),
         );
       },
     );
+  }
+
+  static PersistentBottomSheetController<T> showAsBottomSheet<T>(
+      BuildContext context,
+      {List<T> items,
+      String label,
+      isMenuMode,
+      bool isFilteredOnline,
+      T selectedValue,
+      bool showSearchBox,
+      String dialogTitle,
+      Future<List<T>> Function(String text) onFind,
+      String Function(T item) itemAsString,
+      bool Function(T item, String filter) filterFn,
+      SelectOneItemBuilderType<T> itemBuilder,
+      void Function(T) onChange,
+      InputDecoration searchBoxDecoration,
+      Color backgroundColor,
+      String hintText,
+      double maxHeight,
+      TextStyle dialogTitleStyle}) {
+    return showBottomSheet<T>(
+        context: context,
+        builder: (context) {
+          return Container(
+              height: maxHeight ?? 350,
+              width: double.infinity,
+              padding: EdgeInsets.all(8),
+              decoration: BoxDecoration(
+                color: backgroundColor,
+                borderRadius: BorderRadius.only(
+                    topLeft: Radius.circular(8), topRight: Radius.circular(8)),
+                border: Border.all(color: Colors.blue, width: 2),
+              ),
+              child: SelectDialog<T>(
+                  dialogTitle: dialogTitle,
+                  isMenuMode: isMenuMode,
+                  itemAsString: itemAsString,
+                  filterFn: filterFn,
+                  isFilteredOnline: isFilteredOnline,
+                  selectedValue: selectedValue,
+                  itemsList: items,
+                  onChange: onChange,
+                  maxHeight: maxHeight,
+                  onFind: onFind,
+                  showSearchBox: showSearchBox,
+                  itemBuilder: itemBuilder,
+                  searchBoxDecoration: searchBoxDecoration,
+                  backgroundColor: backgroundColor,
+                  hintText: hintText,
+                  dialogTitleStyle: dialogTitleStyle));
+        });
   }
 
   @override
@@ -81,6 +146,7 @@ class SelectDialog<T> extends StatefulWidget {
 
 class _SelectDialogState<T> extends State<SelectDialog<T>> {
   StreamController<List<T>> itemsStream = StreamController();
+
   final List<T> _items = List();
   final _debouncer = Debouncer();
 
@@ -88,7 +154,9 @@ class _SelectDialogState<T> extends State<SelectDialog<T>> {
   void initState() {
     super.initState();
     Future.delayed(Duration.zero, () async {
-      if (widget.onFind != null) _items.addAll(await widget.onFind(""));
+      if (widget.onFind != null)
+        _items.addAll(
+            await widget.onFind("").catchError((e) => itemsStream.addError(e)));
       if (widget.itemsList != null) _items.addAll(widget.itemsList);
 
       itemsStream.add(_items);
@@ -104,64 +172,29 @@ class _SelectDialogState<T> extends State<SelectDialog<T>> {
   @override
   Widget build(BuildContext context) {
     return Container(
-      width: MediaQuery.of(context).size.width * .9,
-      height: MediaQuery.of(context).size.height * .7,
+      width: MediaQuery.of(context).size.height * .9,
+      height: widget.maxHeight ?? MediaQuery.of(context).size.height * .7,
       child: Column(
         children: <Widget>[
-          if (widget.showSearchBox ?? true)
-            Padding(
-              padding: const EdgeInsets.all(8.0),
-              child: TextField(
-                onChanged: (f) => _debouncer(() {
-                  _onTextChanged(f);
-                }),
-                decoration: widget.searchBoxDecoration ??
-                    InputDecoration(
-                      hintText: widget.hintText,
-                      contentPadding: const EdgeInsets.all(2.0),
-                    ),
-              ),
-            ),
+          _searchField(),
           Expanded(
-            child: Scrollbar(
-              child: StreamBuilder<List<T>>(
-                stream: itemsStream.stream,
-                builder: (context, snapshot) {
-                  if (snapshot.hasError)
-                    return Center(
-                        child: Text("erreur: ${snapshot?.error.toString()}"));
-                  else if (!snapshot.hasData)
-                    return Center(child: CircularProgressIndicator());
-                  else if (snapshot.data.isEmpty)
-                    return Center(child: Text("No data found"));
-                  return ListView.builder(
-                    itemCount: snapshot.data.length,
-                    itemBuilder: (context, index) {
-                      var item = snapshot.data[index];
-                      if (widget.itemBuilder != null)
-                        return InkWell(
-                          child: widget.itemBuilder(
-                              context, item, item == widget.selectedValue),
-                          onTap: () {
-                            widget.onChange(item);
-                            Navigator.pop(context);
-                          },
-                        );
-                      else
-                        return ListTile(
-                          title: Text(widget.itemAsString != null
-                              ? widget.itemAsString(item)
-                              : item.toString()),
-                          selected: item == widget.selectedValue,
-                          onTap: () {
-                            widget.onChange(item);
-                            Navigator.pop(context);
-                          },
-                        );
-                    },
-                  );
-                },
-              ),
+            child: StreamBuilder<List<T>>(
+              stream: itemsStream.stream,
+              builder: (context, snapshot) {
+                if (snapshot.hasError)
+                  return Center(child: Text(snapshot?.error?.toString()));
+                else if (!snapshot.hasData)
+                  return Center(child: CircularProgressIndicator());
+                else if (snapshot.data.isEmpty)
+                  return Center(child: Text("No data found"));
+                return ListView.builder(
+                  itemCount: snapshot.data.length,
+                  itemBuilder: (context, index) {
+                    var item = snapshot.data[index];
+                    return _itemWidget(item);
+                  },
+                );
+              },
             ),
           ),
         ],
@@ -170,8 +203,10 @@ class _SelectDialogState<T> extends State<SelectDialog<T>> {
   }
 
   void _onTextChanged(String filter) async {
-    if (widget.onFind != null && widget.isFilterOnline) {
-      List<T> onlineItems = await widget.onFind(filter);
+    if (widget.onFind != null && widget.isFilteredOnline) {
+      List<T> onlineItems = await widget
+          .onFind(filter)
+          .catchError((e) => itemsStream.addError(e));
       if (onlineItems != null && onlineItems.isNotEmpty) {
         //Remove all old data
         _items.clear();
@@ -183,7 +218,9 @@ class _SelectDialogState<T> extends State<SelectDialog<T>> {
     }
 
     itemsStream.add(_items.where((i) {
-      if (widget.itemAsString != null) {
+      if (widget.filterFn != null)
+        return (widget.filterFn(i, filter));
+      else if (widget.itemAsString != null) {
         return (widget.itemAsString(i))
                 ?.toLowerCase()
                 ?.contains(filter.toLowerCase()) ??
@@ -191,6 +228,49 @@ class _SelectDialogState<T> extends State<SelectDialog<T>> {
       }
       return i.toString().toLowerCase().contains(filter.toLowerCase());
     }).toList());
+  }
+
+  Widget _itemWidget(T item) {
+    if (widget.itemBuilder != null)
+      return InkWell(
+        child: widget.itemBuilder(context, item, item == widget.selectedValue),
+        onTap: () {
+          widget.onChange(item);
+          if (!widget.isMenuMode) Navigator.pop(context);
+        },
+      );
+    else
+      return ListTile(
+        title: Text(widget.itemAsString != null
+            ? widget.itemAsString(item)
+            : item.toString()),
+        selected: item == widget.selectedValue,
+        onTap: () {
+          widget.onChange(item);
+          if (!widget.isMenuMode) Navigator.pop(context);
+        },
+      );
+  }
+
+  Widget _searchField() {
+    return Column(mainAxisSize: MainAxisSize.min, children: <Widget>[
+      if (widget.dialogTitle != null)
+        Text(widget.dialogTitle, style: widget.dialogTitleStyle),
+      if (widget.showSearchBox)
+        Padding(
+            padding: const EdgeInsets.all(8.0),
+            child: TextField(
+              onChanged: (f) => _debouncer(() {
+                _onTextChanged(f);
+              }),
+              decoration: widget.searchBoxDecoration ??
+                  InputDecoration(
+                    hintText: widget.hintText,
+                    border: OutlineInputBorder(),
+                    contentPadding: const EdgeInsets.symmetric(horizontal: 16),
+                  ),
+            ))
+    ]);
   }
 }